requirements.txt
pyscrypt==1.6.2
tinyec==0.3.1
pwinput==1.0.2
GitPython==3.1.24
pycryptodome==3.10.4
secrets==1.0.2
